+++ packages/dataviews/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - DataForm: Add summary field support for both card and panel layouts. The `summary` property has been moved from the field level to the layout level. Fields using `summary` at the field level must now configure it within the `layout` object. [#71576](https://github.com/WordPress/gutenberg/pull/71576)
 
+## 9.1.0 (2025-10-01)
+
 ### Features
 
 - Introduce a new `DataViewsPicker` component. [#70971](https://github.com/WordPress/gutenberg/pull/70971) and [#71836](https://github.com/WordPress/gutenberg/pull/71836).
@@ -20,6 +22,7 @@
 
 - DataViews: Require at least one field to be visible. ([#71625](https://github.com/WordPress/gutenberg/pull/71625))
 - DataViews: Expose `DataViews.FiltersToggled` component to be used in free composition. [#71907](https://github.com/WordPress/gutenberg/pull/71907)
+- DataViews: Add `number` field and refactor `integer` field based on the `number` field. ([#71797](https://github.com/WordPress/gutenberg/pull/71797))
 
 ## 9.0.0 (2025-09-17)